pyproject.toml
[tool.poetry]
name = "sklad"
version = "0.1.0"
description = ""
authors = ["Nachtalb <na@nachtalb.io>"]
readme = "README.md"
license = "LGPL-3.0"

[tool.poetry.dependencies]
python = "^3.12"
twikit = "^1.7.4"
python-telegram-bot = { extras = ["all"], version = "^21.3" }
peewee = "^3.17.5"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ipdb = "^0.13.13"
mypy = "^1.10.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"
ruff-lsp = "^0.0.53"

types-tabulate = "^0.9.0.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
preview = true
target-version = ["py312"]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.12"
show_error_codes = true
pretty = true
strict = true

[[tool.mypy.overrides]]
module = 'tests'
no_warn_no_return = true

[[tool.mypy.overrides]]
module = 'peewee'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'playhouse.shortcuts'
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = 'playhouse.sqlite_ext'
ignore_missing_imports = true

[tool.poetry.scripts]
sklad = "sklad.cli:main"
sklad-bot = "sklad.main:main"
